# post/docs/swagger.yaml
basePath: /
definitions:
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.Post:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
  main.PostCreate:
    properties:
      body:
        type: string
    type: object
  main.PostUpdate:
    properties:
      body:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Handles creating, reading, updating, and deleting posts.
  title: Post Service API
  version: "3.0"
paths:
  /posts:
    get:
      description: Returns every post
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Creates a new post for the requesting user
      parameters:
      - description: Requesting user's ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Post body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/main.PostCreate'
      produces:
      - application/json
      responses:
        "200":
          description: post created
        "400":
          description: missing/malformed body
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: missing user id
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a post
      tags:
      - posts
  /posts/{post_id}:
    delete:
      description: Deletes a post owned by the requesting user
      parameters:
      - description: Requesting user's ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      responses:
        "200":
          description: post deleted
        "400":
          description: missing post id
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: missing user id
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a post
      tags:
      - posts
    get:
      description: Returns one post by ID
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Post'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get a single post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Updates an existing post's body
      parameters:
      - description: Requesting user's ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Updated post body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/main.PostUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: post updated
        "400":
          description: missing post id or malformed body
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: missing user id
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Update a post
      tags:
      - posts
swagger: "2.0"